+++ lib/cinder_web/live/series_detail_live.ex
@@ -225,7 +225,7 @@ defmodule CinderWeb.SeriesDetailLive do
   end
 
   def handle_event("set_series_language", %{"preferred_language" => lang}, socket)
-      when lang in ["original", "french", "any"] do
+      when lang in ["original", "french", "dual", "any"] do
     case Catalog.set_series_language(socket.assigns.series, lang) do
       {:ok, series} ->
         {:noreply, assign(socket, :series, series)}
@@ -244,19 +244,6 @@ defmodule CinderWeb.SeriesDetailLive do
 
       {:error, _} ->
         {:noreply, put_flash(socket, :error, gettext("Couldn't update the profile."))}
-    end
-  end
-
-  def handle_event("set_anime_audio_mode", %{"anime_audio_mode" => mode}, socket)
-      when mode in ["", "original", "dub", "dual", "any"] do
-    value = if mode == "", do: nil, else: String.to_existing_atom(mode)
-
-    case Catalog.set_anime_audio_mode(socket.assigns.series, value) do
-      {:ok, _} ->
-        {:noreply, reload(socket)}
-
-      {:error, _} ->
-        {:noreply, put_flash(socket, :error, gettext("Couldn't update the audio mode."))}
     end
   end
 
@@ -675,14 +662,6 @@ defmodule CinderWeb.SeriesDetailLive do
             <.profile_select field={@profile_form[:media_profile]} />
           </.form>
           <.profile_summary id="series-profile-summary" summary={@profile_summary} />
-          <form
-            :if={@profile_summary.effective == :anime}
-            id="series-anime-audio-form"
-            phx-change="set_anime_audio_mode"
-            class="mt-2"
-          >
-            <.anime_audio_mode_select value={@series.anime_audio_mode} />
-          </form>
         </div>
       </div>